# swagger-export.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/wallet": {
      "get": {
        "description": "Retrieve complete wallet information including TechCoin balance, XP, and recent transactions",
        "operationId": "WalletController_getWallet",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Wallet information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 1,
                    "userId": 123,
                    "techCoin": 250,
                    "xp": 850,
                    "createdAt": "2025-07-16T00:00:00Z",
                    "updatedAt": "2025-07-16T12:00:00Z",
                    "recentTransactions": [
                      {
                        "id": 5,
                        "type": "EARN",
                        "amount": 25,
                        "description": "Task completion reward",
                        "createdAt": "2025-07-16T10:30:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user wallet information",
        "tags": [
          "💰 Digital Wallet & TechCoin"
        ]
      }
    },
    "/wallet/buy": {
      "post": {
        "description": "Buy TechCoin using mock Stripe payment integration. 1 USD = 10 TechCoin",
        "operationId": "WalletController_buyTechCoin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyTechCoinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "TechCoin purchased successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "transaction": {
                      "id": 10,
                      "type": "BUY",
                      "amount": 100,
                      "description": "Purchased 100 TechCoin for $10.00",
                      "stripePaymentId": "pi_mock_12345",
                      "createdAt": "2025-07-16T12:30:00Z"
                    },
                    "wallet": {
                      "techCoin": 350,
                      "xp": 850
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Payment failed or invalid amount"
          },
          "401": {
            "description": "Unauthorized - JWT token required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Purchase TechCoin",
        "tags": [
          "💰 Digital Wallet & TechCoin"
        ]
      }
    },
    "/wallet/spend": {
      "post": {
        "description": "Spend TechCoin for various platform activities (joining challenges, creating cafes, etc.)",
        "operationId": "WalletController_spendTechCoin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpendTechCoinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "TechCoin spent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "transaction": {
                      "id": 11,
                      "type": "SPEND",
                      "amount": -50,
                      "description": "Joined challenge: Code Master",
                      "createdAt": "2025-07-16T12:45:00Z"
                    },
                    "remainingBalance": 300
                  }
                }
              }
            }
          },
          "400": {
            "description": "Insufficient TechCoin balance"
          },
          "401": {
            "description": "Unauthorized - JWT token required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Spend TechCoin",
        "tags": [
          "💰 Digital Wallet & TechCoin"
        ]
      }
    },
    "/wallet/earn": {
      "post": {
        "description": "Earn TechCoin through various activities (completing tasks, winning challenges, etc.)",
        "operationId": "WalletController_earnTechCoin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EarnTechCoinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "TechCoin earned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "transaction": {
                      "id": 12,
                      "type": "EARN",
                      "amount": 75,
                      "description": "Won challenge: Code Master",
                      "createdAt": "2025-07-16T13:00:00Z"
                    },
                    "newBalance": 375,
                    "xpGained": 25
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Earn TechCoin",
        "tags": [
          "💰 Digital Wallet & TechCoin"
        ]
      }
    },
    "/wallet/transactions": {
      "get": {
        "description": "Retrieve paginated transaction history for the user wallet",
        "operationId": "WalletController_getTransactionHistory",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by transaction type",
            "schema": {
              "enum": [
                "BUY",
                "SPEND",
                "EARN"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "transactions": [
                      {
                        "id": 12,
                        "type": "EARN",
                        "amount": 75,
                        "description": "Won challenge: Code Master",
                        "createdAt": "2025-07-16T13:00:00Z"
                      },
                      {
                        "id": 11,
                        "type": "SPEND",
                        "amount": -50,
                        "description": "Joined challenge: Code Master",
                        "createdAt": "2025-07-16T12:45:00Z"
                      }
                    ],
                    "pagination": {
                      "page": 1,
                      "limit": 20,
                      "total": 25,
                      "totalPages": 2
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transaction history",
        "tags": [
          "💰 Digital Wallet & TechCoin"
        ]
      }
    },
    "/wallet/balance": {
      "get": {
        "description": "Get current TechCoin balance and XP, optionally check if user has enough for a specific amount",
        "operationId": "WalletController_checkBalance",
        "parameters": [
          {
            "name": "amount",
            "required": false,
            "in": "query",
            "description": "Amount to check if user has enough TechCoin",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balance information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "balance": 375,
                    "xp": 875,
                    "hasEnough": true,
                    "requiredAmount": 100
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check wallet balance",
        "tags": [
          "💰 Digital Wallet & TechCoin"
        ]
      }
    },
...
  },
  "info": {
    "title": "TechVerse Café API",
    "description": "Backend API for TechVerse Café MVP - A tech community platform with digital economy\n\n## 🚀 New Features in MVP\n- � **Digital Wallet & TechCoin System**: Earn, buy, and spend TechCoin\n- � **Challenges & Competitions**: Community challenges with rewards\n- �️ **Enhanced Projects System**: Task management with payments\n- 🌍 **Multi-language Support**: Localized content and UI\n- ☕ **Café Communities**: Topic-based discussion groups\n\n## Core Features\n- � JWT Authentication & Role-based Access Control\n- 📝 Content Management (Articles, Posts)\n- 🔖 Bookmarking System\n- 👥 User Profiles & Social Features\n- 🤖 AI Integration for content generation\n\n## Authentication\nMost endpoints require authentication using JWT tokens. Include the token in the `X-HTTP-TOKEN` header.\n\n## Role-Based Access Control\nThe platform uses a two-tier role system:\n- **Global Roles**: ADMIN, SUPERVISOR, EDITOR, MEMBER\n- **Café Roles**: BARISTA, THINKER, JOURNALIST, MEMBER\n\n## Getting Started\n1. Register a new user account\n2. Login to receive JWT token\n3. Use the token in `X-HTTP-TOKEN` header for authenticated requests\n4. Create your wallet and start earning TechCoin\n5. Join challenges and participate in projects\n\n## Download API Specification\n- JSON Format: [/api/swagger.json](/api/swagger.json)\n- YAML Format: [/api/swagger.yaml](/api/swagger.yaml)\n      ",
    "version": "2.0.0",
...
    },
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "X-HTTP-TOKEN": {
        "type": "apiKey",
        "in": "header",
        "name": "X-HTTP-TOKEN",
        "description": "JWT token for authentication. Format: Bearer <token>"
      }
    },
    "schemas": {
...
      "BuyTechCoinDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Amount of TechCoin to purchase",
            "example": 100,
            "minimum": 1,
            "maximum": 10000
          },
          "paymentMethod": {
            "type": "string",
            "description": "Payment method (mock for demo)",
            "example": "stripe",
            "default": "stripe"
          }
        },
        "required": [
          "amount"
        ]
      },
      "SpendTechCoinDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Amount of TechCoin to spend",
            "example": 50,
            "minimum": 1
          },
          "description": {
            "type": "string",
            "description": "Description of the spending transaction",
            "example": "Join challenge: Code Master",
            "maxLength": 255
          },
          "referenceId": {
            "type": "number",
            "description": "Reference ID for the transaction (challenge ID, cafe ID, etc.)",
            "example": 123
          },
          "category": {
            "type": "string",
            "description": "Type of spending activity",
            "example": "challenge_entry",
            "enum": [
              "challenge_entry",
              "cafe_creation",
              "task_escrow",
              "other"
            ]
          }
        },
        "required": [
          "amount",
          "description"
        ]
      },
      "EarnTechCoinDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Amount of TechCoin earned",
            "example": 25,
            "minimum": 1
          },
          "description": {
            "type": "string",
            "description": "Description of how the TechCoin was earned",
            "example": "Completed task: Fix login bug",
            "maxLength": 255
          },
          "xpReward": {
            "type": "number",
            "description": "XP points to award along with TechCoin",
            "example": 10,
            "default": 0
          },
          "referenceId": {
            "type": "number",
            "description": "Reference ID for the earning activity",
            "example": 456
          },
          "category": {
            "type": "string",
            "description": "Category of earning activity",
            "example": "task_completion",
            "enum": [
              "task_completion",
              "challenge_win",
              "bonus_reward",
              "other"
            ]
          }
        },
        "required": [
          "amount",
          "description"
        ]
      },
...
    }
  }
}